Size Constraint has Very Important, might need Required I included the most threatening thing at the top.  Size constraint before wasn't actually on the list of constraints, and was assumed to be included based on the number of panelists (I believe).  Making it required results in about 20% of our panels being unscheduled, so it's made as Very Important for now (where it makes maybe five more panels unschedulable).
Change Log
* Added Size Constraint addition to Panel Constructor
* Changed minSize in SizeConstraint from short to int (finally)
* Completed ConstraintFactory (for the most part)
** Fixed PanelistConstraint being created as CategoryConstraint
** Added MaxPanelist, MinPanelist (as our heuristic priority 3, not priority 1; it didn't seem to have issues with either though)
* Added guard conditions against null pointers in MaxPanelist and MinPanelist

To-Do
* Add PanelistAvaliabilityConstraint (requires construction from parser, as it isn't in the Constraints list)

I think this should fix a lot of the bugs we've been seeing.

+++ src/symposium/model/SizeConstraint.java
@@ -2,13 +2,13 @@ package symposium.model;
 
 public class SizeConstraint extends Constraint {
 
-    short minSize;
+    int minSize;
 
     /**
      *
      * @param mSize The minimum size necessary for the panel to fit.
      */
-    public SizeConstraint(ConstraintPriority priority, Panel p, short mSize) {
+    public SizeConstraint(ConstraintPriority priority, Panel p, int mSize) {
         super(priority, p);
         minSize = mSize;
     }
@@ -32,6 +32,6 @@ public class SizeConstraint extends Constraint {
     }
 
     public int getMinSize(){
-        return (int) this.minSize;
+        return this.minSize;
     }
 }
